services/ecommerce/pyproject.toml
[tool.poetry]
name = "grpc-ecommerce-service"
version = "0.1.0"
description = ""
authors = ["FernandoPrzGmz"]
readme = "README.md"
packages = [{include = "grpc_ecommerce_service"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.2.9"
mysqlclient = "2.2.1"
djangorestframework = "3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
